Hackathon/AttentionX/requirements.txt
streamlit
moviepy
librosa
opencv-python
mediapipe
openai-whisper
numpy
